fix: solve garbage collection issues
This fix makes it so that objects such as a `GameObject` can be created
from Lua without causing memory corruption bugs.

It also fixes an issue with the camera where zoom per frame was not
multiplied by the deltatime causing zooming with the camera to be very
difficult.

Also added more example Lua API usage.